requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
google-generativeai